move services.coordinator to services
